Implementation for run-image link and single-hop
This uses a stamp file in 'lib/modules' jimage in order to recognize
multi-hop links. However, this has the caveat of no longer producing
reproducible builds as compared to a packaged module-based link.

Add an option to allow for multi-hop, which disables the stamp file
addition and makes it reproducible again.