The constants such as the pathname of the timestamp file and the internal file listing per-module non-class non-resource files are part of jlink.  Move the constants to JlinkTask to follow where OPTIONS_RESOURCE is defined.
JRTArchive scans the class and resource files of a given module from the runtime image.
It should also read `fs_$MODULE_files` to find the list of non-class and non-resource files.

The current implementation checks if a file is modified lazily when Entry::stream
is called and so it has to remember if file modification has been checked and
warning has been emitted.   I think doing the file modification check eagerly
when collectFiles is called would simplify the code.

For maintainability, better to move the reading of and writing to `fs_$MODULE_files`
together in a single class rather than separated in JRTArchive and JlinkResourcesListPlugin.