Fix provider verification when some JMODs are present
In some configurations, e.g. when java.base is missing from the packaged
modules, but another JDK module is present as JMOD that is linked into
an image, then provider verification can fail. Thus, the run-time image
link fails. Verify that this doesn't happen.

The fix is to return Platform.runtime() for run-time image based links
as well. Otherwise this code would return the wrong result.