Add /othervm to some langtools tier1 tests
Those tests are using module patches from JTREG. Since the run-time
image based link uses ModuleFinder.ofSystem(), which will see the extra
classes comming from the module patch. Then the size look-up using the
JRT FS provider fails, since that only looks in the module image
(lib/modules) and NOT also in the patch. Thus, we get a
NoSuchFileException and the link fails.

Run the tests with othervm so that the JTREG patch'ed module isn't
visible to the test.